1.ゲームオブジェクトの位置・回転・スケールをトランスフォームで管理するように変更、2.namespace 「Runtime」を廃止
[Update]

// 不足する処理の追加
Transform

// Runtime の廃止により修正
Shader
Material

// トランスフォームの採用＆Runtimeの廃止により修正
CameraController_3D
CameraControllPanel
ModelControlPanel

Mesh
Model
Sprite
GameObject
Camera

+++ Source/Framework/Render/Resource/Texture.cpp
@@ -84,9 +84,9 @@ bool Render::Texture::Init(const ImageData& _imageData)
 	result = m_pResource->WriteToSubresource(
 		0,
 		nullptr,
-		image->pixels,		
-		image->rowPitch,	// 1ラインサイズ
-		image->slicePitch	// 1枚サイズ
+		image->pixels,
+		(UINT)image->rowPitch,	// 1ラインサイズ
+		(UINT)image->slicePitch	// 1枚サイズ
 	);
 
 	if (FAILED(result))
